requirements.txt
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.20
django-allauth==0.35.0
django-crispy-forms==1.7.2
djangorestframework==3.8.2
graphviz==0.8.3
gunicorn==19.8.1
idna==2.6
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
oauthlib==2.0.7
pandas==0.23.0
Pillow==5.1.0
psycopg2
pydotplus==2.0.2
pyparsing==2.2.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
requests==2.20.0
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
six==1.11.0
urllib3==1.24.2
